requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=2.0.0
pydantic-settings>=2.0.0
loguru>=0.6.0
google-cloud-speech>=2.0.0
google-cloud-storage>=2.0.0
google-api-python-client>=2.0.0
vertexai>=0.0.1
ffmpeg-python>=0.2.0
numpy>=1.21.0
requests>=2.26.0
python-jose>=3.3.0
websockets>=10.0
google-cloud-language>=2.0.0
aiofiles>=0.8.0
python-dotenv>=0.19.0
pytz>=2024.2
